fix printf format for time_t.  notified by casha@e7.pl.  henning ok

+++ libexec/spamd/spamd.c
@@ -1,4 +1,4 @@
-/*	$OpenBSD: spamd.c,v 1.35 2003/06/11 14:24:46 deraadt Exp $	*/
+/*	$OpenBSD: spamd.c,v 1.36 2003/07/29 18:39:23 deraadt Exp $	*/
 
 /*
  * Copyright (c) 2002 Theo de Raadt.  All rights reserved.
@@ -534,7 +534,8 @@ closecon(struct con *cp)
 		time_t t;
 
 		time(&t);
-		printf("%s connected for %d seconds.\n", cp->addr, t - cp->s);
+		printf("%s connected for %ld seconds.\n", cp->addr,
+		    (long)(t - cp->s));
 	}
 	if (cp->lists != NULL) {
 		free(cp->lists);
